jpa-for-h2

+++ src/main/java/com/ex/repository/MemberRepository.java
@@ -1,5 +1,6 @@
 package com.ex.repository;
 
+
 import com.ex.domain.Member;
 
 import java.util.List;
@@ -7,8 +8,7 @@ import java.util.Optional;
 
 public interface MemberRepository {
   Member insert(Member member);
-  Optional<Member> findById(Long id);
+  Optional<Member> findById(int id);
   Optional<Member> findByName(String name);
   List<Member> findAll();
-
 }
